Updated example

+++ examples/strict_class_example.py
@@ -2,7 +2,7 @@ from pyparcel import Long, UnsignedLongLong, Double
 
 
 class Foo:
-    def __init__(self, a: int, b: int, c: int, d: int):
+    def __init__(self, a: int, b: int, c: int, d: long):
         self.a = a
         self.b = Long(b)
         self.c = UnsignedLongLong(c)
